split RDP file loading
